+++ src/HetrickCV.hpp
@@ -32,6 +32,7 @@ extern Model *modelTwoToFour;
 extern Model *modelAnalogToDigital;
 extern Model *modelASR;
 extern Model *modelBinaryGate;
+extern Model *modelBinaryNoise;
 extern Model *modelBitshift;
 extern Model *modelBlankPanel;
 extern Model *modelBoolean3;